+++ src/App.js
@@ -1,12 +1,22 @@
 import React from "react";
 import { Header } from "./components/Header";
+import { Balance } from "./components/Balance";
+import { IncExpenses } from "./components/IncExpenses";
+import { TransactionList } from "./components/TransactionList";
+import { AddTransaction } from "./components/AddTransaction";
 
 import "./App.css"
 
 function App() {
   return (
-    <div className="App">
-    <Header />
+    <div>
+      <Header />
+      <div className="container">
+        <Balance />
+        <IncExpenses />
+        <TransactionList />
+        <AddTransaction />
+      </div>
     </div>
   );
 }